rework how incidents pass info to raid workers, and add surprise logging
There does not seem to be a simple way to decide something in the incident
worker and than pass that to the raid worker, which I guess is why
poaching was copy&pasting TryExecuteWorker(). But that feels rather fragile,
so instead use a (hopefully) less fragile hack of using raid worker's
SpawnThreats(), which is called at a suitable place from incident worker.